chore: extract reverse proxy from pkg/gateway into pkg/edge (Phase 2)
Closes the original Phase 2 from #37. With observability primitives now in
pkg/observability (#44), the L7 reverse proxy and its TLS-cert helper can
finally leave pkg/gateway without circular imports.

Moved verbatim into pkg/edge:
- proxy.go: ProxyRoute, ProxyServer, ServicePorts, BuildRoutes,
  BuildRoutesWithPorts, NewProxyServer{,WithOpts}, StartProxy{,WithOpts},
  ProxyOpts, plus all internal handlers (logProxyRequest,
  proxyToWithOpts, addCORSHeaders, rewriteResponseBody,
  isWebSocketUpgrade, proxyWebSocket, statusRecorder).
- certs.go: CertPair, EnsureCerts, all OS trust helpers, file/PEM I/O.
- proxy_test.go and certs_test.go.

Internal references to RequestLog / RequestStats / RequestBroadcaster /
RequestEntry / GenerateTraceID updated to import from pkg/observability
directly (no longer go through gateway aliases).

pkg/gateway/proxy.go and pkg/gateway/certs.go shrink to type aliases +
forwarder functions so external callers (cmd/gateway/main.go and any
historical importers) continue to compile.

cmd/gateway/main.go migrated to import pkg/edge directly — edge is the
canonical home now. The gateway aliases stay in place for safety.

Test sweep green except the same pre-existing pkg/platform/store
testcontainers/Docker failure that's been failing on every PR in this
branch (Docker not running on this machine; reproduces on origin/main).

Closes #45
