+++ pkg/edge/certs_test.go
@@ -1,4 +1,4 @@
-package gateway
+package edge
 
 import (
 	"testing"
